move graph/search to engine (#4596)
