+++ extra/optimization/get_action_space.py
@@ -1,7 +1,7 @@
 import random
 from tqdm import tqdm
 from extra.optimization.helpers import load_worlds, ast_str_to_lin
-from tinygrad.features.search import actions
+from tinygrad.engine.search import actions
 from tinygrad.codegen.linearizer import Linearizer
 
 tactions = set()
