+++ extra/optimization/test_time_linearizer.py
@@ -1,5 +1,5 @@
 from extra.optimization.helpers import load_worlds, ast_str_to_lin
-from tinygrad.features.search import bufs_from_lin, time_linearizer, get_linearizer_actions
+from tinygrad.engine.search import bufs_from_lin, time_linearizer, get_linearizer_actions
 
 if __name__ == "__main__":
   ast_strs = load_worlds()
