+++ test/external/speed_beam_v_hcopt.py
@@ -1,6 +1,6 @@
 from tinygrad import Device
 from tinygrad.helpers import getenv, DEBUG, BEAM
-from tinygrad.features.search import beam_search, time_linearizer, bufs_from_lin
+from tinygrad.engine.search import beam_search, time_linearizer, bufs_from_lin
 from extra.optimization.helpers import load_worlds, ast_str_to_lin
 
 if __name__ == "__main__":
